+++ packages/ses/test/console-error-trap/exit-code.js
@@ -1,0 +1,3 @@
+import '../../index.js';
+import './exit-code-lockdown.js';
+import './hazard.js';