+++ packages/ses/test/console-error-trap/report-lockdown.js
@@ -1,0 +1,1 @@
+lockdown({ errorTrapping: 'report' });